+++ CHANGES.md
@@ -1,5 +1,6 @@
 ## Version next
 
+- fix(properties) fix incorrect handling of non-alphanumeric keys [Egor Rogov][]
 - enh(parser) Detect comments based on english like text, rather than keyword list [Josh Goebel][]
 - enh(shell) add alias ShellSession [Ryan Mulligan][]
 - enh(shell) consider one space after prompt as part of prompt [Ryan Mulligan][]